refactor(tests): Refactor tests, split them to avoid too large files
